+++ src/main/java/controller/JanggiController.java
@@ -8,11 +8,11 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
-import model.Board;
-import model.BoardFactory;
 import model.JanggiGame;
 import model.Team;
-import model.coordinate.Position;
+import model.board.Board;
+import model.board.BoardFactory;
+import model.board.Position;
 import model.formation.FormationFactory;
 import model.formation.JanggiFormation;
 import model.piece.Piece;
