+++ src/main/java/model/board/Board.java
@@ -1,10 +1,9 @@
-package model;
+package model.board;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import model.coordinate.Position;
 import model.piece.Piece;
 
 public class Board {
