+++ src/main/java/model/formation/MaSangSangMaStrategy.java
@@ -1,21 +1,20 @@
 package model.formation;
 
+import static model.board.Position.CHO_LEFT_INNER;
+import static model.board.Position.CHO_LEFT_OUTER;
+import static model.board.Position.CHO_RIGHT_INNER;
+import static model.board.Position.CHO_RIGHT_OUTER;
+import static model.board.Position.HAN_LEFT_INNER;
+import static model.board.Position.HAN_LEFT_OUTER;
+import static model.board.Position.HAN_RIGHT_INNER;
+import static model.board.Position.HAN_RIGHT_OUTER;
+
+import java.util.Map;
 import model.Team;
-import model.coordinate.Position;
+import model.board.Position;
 import model.piece.Elephant;
 import model.piece.Horse;
 import model.piece.Piece;
-
-import java.util.Map;
-
-import static model.coordinate.Position.CHO_LEFT_INNER;
-import static model.coordinate.Position.CHO_LEFT_OUTER;
-import static model.coordinate.Position.CHO_RIGHT_INNER;
-import static model.coordinate.Position.CHO_RIGHT_OUTER;
-import static model.coordinate.Position.HAN_LEFT_INNER;
-import static model.coordinate.Position.HAN_LEFT_OUTER;
-import static model.coordinate.Position.HAN_RIGHT_INNER;
-import static model.coordinate.Position.HAN_RIGHT_OUTER;
 
 public class MaSangSangMaStrategy extends FormationStrategy {
     @Override
