+++ src/main/java/model/formation/SangMaSangMaStrategy.java
@@ -1,21 +1,20 @@
 package model.formation;
 
+import static model.board.Position.CHO_LEFT_INNER;
+import static model.board.Position.CHO_LEFT_OUTER;
+import static model.board.Position.CHO_RIGHT_INNER;
+import static model.board.Position.CHO_RIGHT_OUTER;
+import static model.board.Position.HAN_LEFT_INNER;
+import static model.board.Position.HAN_LEFT_OUTER;
+import static model.board.Position.HAN_RIGHT_INNER;
+import static model.board.Position.HAN_RIGHT_OUTER;
+
+import java.util.Map;
 import model.Team;
-import model.coordinate.Position;
+import model.board.Position;
 import model.piece.Elephant;
 import model.piece.Horse;
 import model.piece.Piece;
-
-import java.util.Map;
-
-import static model.coordinate.Position.CHO_LEFT_INNER;
-import static model.coordinate.Position.CHO_LEFT_OUTER;
-import static model.coordinate.Position.CHO_RIGHT_INNER;
-import static model.coordinate.Position.CHO_RIGHT_OUTER;
-import static model.coordinate.Position.HAN_LEFT_INNER;
-import static model.coordinate.Position.HAN_LEFT_OUTER;
-import static model.coordinate.Position.HAN_RIGHT_INNER;
-import static model.coordinate.Position.HAN_RIGHT_OUTER;
 
 public class SangMaSangMaStrategy extends FormationStrategy {
     @Override
