+++ src/main/java/model/piece/Piece.java
@@ -2,7 +2,7 @@ package model.piece;
 
 import java.util.List;
 import model.Team;
-import model.coordinate.Position;
+import model.board.Position;
 
 public abstract class Piece {
