+++ src/main/java/view/OutputView.java
@@ -1,11 +1,5 @@
 package view;
 
-import model.Board;
-import model.coordinate.Position;
-import model.piece.Piece;
-
-import java.util.Map;
-
 import static view.formater.BoardFormatter.COL_NUM;
 import static view.formater.BoardFormatter.RED;
 import static view.formater.BoardFormatter.RESET;
@@ -15,17 +9,12 @@ import static view.formater.BoardFormatter.VERTICAL_LINE;
 import static view.formater.BoardFormatter.formatHorizon;
 import static view.formater.BoardFormatter.formatSymbol;
 
+import java.util.Map;
+import model.board.Board;
+import model.board.Position;
+import model.piece.Piece;
+
 public class OutputView {
-
-    public void displayBoard(Map<Position, Piece> board) {
-        displayColIndex();
-        String border = formatHorizon(Board.BOARD_COL);
-        System.out.println(border);
-
-        displayPositionByPiece(board);
-
-        System.out.println(border);
-    }
 
     private static void displayPositionByPiece(Map<Position, Piece> board) {
         for (int row = 0; row < Board.BOARD_ROW; row++) {
@@ -45,6 +34,16 @@ public class OutputView {
             System.out.print(SPACE + column);
         }
         System.out.println();
+    }
+
+    public void displayBoard(Map<Position, Piece> board) {
+        displayColIndex();
+        String border = formatHorizon(Board.BOARD_COL);
+        System.out.println(border);
+
+        displayPositionByPiece(board);
+
+        System.out.println(border);
     }
 
     public void displayError(String message) {
