+++ src/test/java/model/board/DirectionTest.java
@@ -1,12 +1,10 @@
-package model;
-
-import model.coordinate.Direction;
-import model.coordinate.Position;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
+package model.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 class DirectionTest {
