+++ src/test/java/model/formation/FormationFactoryTest.java
@@ -1,15 +1,14 @@
 package model.formation;
 
-import model.coordinate.Position;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.Map;
+import java.util.stream.Stream;
+import model.board.Position;
 import model.piece.Piece;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-
-import java.util.Map;
-import java.util.stream.Stream;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 class FormationFactoryTest {
