+++ src/test/java/model/formation/FormationStrategyTest.java
@@ -1,23 +1,22 @@
 package model.formation;
 
 
+import static model.board.Position.CHO_LEFT_INNER;
+import static model.board.Position.CHO_LEFT_OUTER;
+import static model.board.Position.CHO_RIGHT_INNER;
+import static model.board.Position.CHO_RIGHT_OUTER;
+import static model.board.Position.HAN_LEFT_INNER;
+import static model.board.Position.HAN_LEFT_OUTER;
+import static model.board.Position.HAN_RIGHT_INNER;
+import static model.board.Position.HAN_RIGHT_OUTER;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.Map;
 import model.Team;
-import model.coordinate.Position;
+import model.board.Position;
 import model.piece.Piece;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-
-import java.util.Map;
-
-import static model.coordinate.Position.CHO_LEFT_INNER;
-import static model.coordinate.Position.CHO_LEFT_OUTER;
-import static model.coordinate.Position.CHO_RIGHT_INNER;
-import static model.coordinate.Position.CHO_RIGHT_OUTER;
-import static model.coordinate.Position.HAN_LEFT_INNER;
-import static model.coordinate.Position.HAN_LEFT_OUTER;
-import static model.coordinate.Position.HAN_RIGHT_INNER;
-import static model.coordinate.Position.HAN_RIGHT_OUTER;
-import static org.assertj.core.api.Assertions.assertThat;
 
 class FormationStrategyTest {
