+++ src/test/java/model/testdouble/SpyBoard.java
@@ -1,9 +1,9 @@
 package model.testdouble;
 
 import java.util.Map;
-import model.Board;
 import model.Team;
-import model.coordinate.Position;
+import model.board.Board;
+import model.board.Position;
 import model.piece.Horse;
 import model.piece.Piece;
